+++ application/auth/usecases/SendCodeUseCase.ts
@@ -1,6 +1,21 @@
 import { supabase } from '@/app/shared/lib/supabase';
-import { createVerificationToken } from '@/app/shared/lib/verification-jwt';
 import { EmailService } from '@/domain/auth/services/EmailService';
+import jwt from 'jsonwebtoken';
+
+const SECRET = process.env.JWT_SECRET!;
+
+type VerificationPayload = {
+  email: string;
+  code: string;
+};
+
+export const createVerificationToken = (payload: VerificationPayload) => {
+  return jwt.sign(payload, SECRET, { expiresIn: '5m' });
+};
+
+export const verifyVerificationToken = (token: string): VerificationPayload => {
+  return jwt.verify(token, SECRET) as VerificationPayload;
+};
 
 export class SendCodeUseCase {
   constructor(private readonly emailService: EmailService) {}
